fix : COLORS 상수 colors.ts로 분리
