+++ src/shared/lib/colors.ts
@@ -38,5 +38,14 @@ export const colors = {
   black: '#000000',
 };
 
+export const COLORS = {
+  primary: colors['primary'][300],
+  gray: colors['gray'][300],
+  red: colors['red'][300],
+  green: colors['green'][300],
+  black: colors['black'],
+};
+
 export type Colors = keyof typeof colors;
+export type IconColor = keyof typeof COLORS;
 export const colorNames = Object.keys(colors) as Colors[];
